chore: moved all presentations related pages to the presentation route group
In this way we can apply the presentation layout to only the presentation related pages
