+++ src/app/(presentation)/[presentationSection]/[slideUrl]/page.module.scss
