+++ src/app/(presentation)/page.js
@@ -1,5 +1,5 @@
 "use client";
-import Intro from "../components/Intro/Intro";
+import Intro from "../../components/Intro/Intro";
 
 export default function Home() {
     return (
