--- src/app/[presentationSection]/page.js
@@ -1,7 +1,0 @@
-export default function Page() {
-    return (
-        <div>
-            CIAO 2
-        </div>
-    );
-}
